324128ing

+++ code/ZeroJudge/g012/g012.cpp
@@ -49,7 +49,7 @@ void solve() {
     cin>>n;
     int now=0;
     ll s=0,an=0;
-    vector<int> a(n);
+    vector<ll> a(n);
     REP(i,n) cin>>a[i];
     for(int i=1;i<1<<n;i++) {
         int j=i&-i;
